Allow empty response body
The pattern match to fetch the call response body assumes that the
returned type will have children to parse into a map. In some cases,
APIs will return elements with no children (for example, the
`setSessionParameters` call on the Five9 Statistics API). This change
adds an additional pattern match to handle the case where the element
has no children and returns an empty map.